removing the release2 branch and moving it to new folder
