+++ src/work-allocation-2/components/available-tasks-filter/available-tasks-filter.component.scss
@@ -1,4 +1,4 @@
-:host ::ng-deep .available-tasks_location-filter {
+:host ::ng-deep .available-tasks_location-filter-r2 {
   .xui-checkbox-list {
     height: 200px;
     // border-bottom: solid 1px #b1b4b6;
