+++ src/work-allocation-2/components/available-tasks-filter/available-tasks-filter.component.spec.ts
@@ -8,11 +8,11 @@ import { of } from 'rxjs';
 import * as dtos from '../../models/dtos';
 import { LocationDataService } from '../../services';
 import { getMockLocations } from '../../tests/utils.spec';
-import { FilterConstants } from '../constants';
+import { FilterConstants } from '../../components/constants';
 import { AvailableTasksFilterComponent } from './available-tasks-filter.component';
 
 @Component({
-  template: `<exui-available-tasks-filter (selectionChanged)="onSelectionChanged($event)"></exui-available-tasks-filter>`
+  template: `<exui-available-tasks-filter-r2 (selectionChanged)="onSelectionChanged($event)"></exui-available-tasks-filter-r2>`
 })
 class WrapperComponent {
   @ViewChild(AvailableTasksFilterComponent) public appComponentRef: AvailableTasksFilterComponent;
