+++ src/work-allocation-2/components/available-tasks-filter/available-tasks-filter.component.ts
@@ -6,10 +6,10 @@ import { SessionStorageService } from '../../../app/services';
 import { Location } from '../../models/dtos';
 import { LocationDataService } from '../../services';
 import { handleFatalErrors, WILDCARD_SERVICE_DOWN } from '../../utils';
-import { FilterConstants } from '../constants';
+import { FilterConstants } from '../../components/constants';
 
 @Component({
-  selector: 'exui-available-tasks-filter',
+  selector: 'exui-available-tasks-filter-r2',
   templateUrl: './available-tasks-filter.component.html',
   styleUrls: ['available-tasks-filter.component.scss']
 })
