+++ src/work-allocation-2/components/case-name-field/case-name-field.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, OnChanges } from '@angular/core';
 
-import { AppConstants } from '../../../../app/app.constants';
+import { AppConstants } from '../../../app/app.constants';
 
 @Component({
   selector: 'exui-case-name-field',
