+++ src/work-allocation-2/containers/available-tasks/available-tasks.component.html
@@ -1,17 +1,5 @@
-<ng-container [ngSwitch]="(featureVersion$ | async)">
-  <ng-container *ngSwitchCase="'WorkAllocationRelease2'">
-    <exui-task-list-r2
-      [taskServiceConfig]="taskServiceConfig" [tasks]="tasks" [fields]="fields"
-      [sortedBy]="sortedBy" [emptyMessage]="emptyMessage"
-      (sortEvent)="onSortHandler($event)" (actionEvent)="onActionHandler($event)">
-    </exui-task-list-r2>
-  </ng-container>
-  <ng-container *ngSwitchDefault>
-    <exui-available-tasks-filter (selectionChanged)="onLocationsChanged($event)"></exui-available-tasks-filter>
-    <exui-task-list
-      [taskServiceConfig]="taskServiceConfig" [tasks]="tasks" [fields]="fields"
-      [sortedBy]="sortedBy" [emptyMessage]="emptyMessage"
-      (sortEvent)="onSortHandler($event)" (actionEvent)="onActionHandler($event)">
-    </exui-task-list>
-  </ng-container>
-</ng-container>
+<exui-task-list-r2
+  [taskServiceConfig]="taskServiceConfig" [tasks]="tasks" [fields]="fields"
+  [sortedBy]="sortedBy" [emptyMessage]="emptyMessage"
+  (sortEvent)="onSortHandler($event)" (actionEvent)="onActionHandler($event)">
+</exui-task-list-r2>
