+++ src/work-allocation-2/containers/task-list/task-list.component.html
@@ -1,3 +1,4 @@
+<p *ngIf="tasks">Displaying <span style="font-weight:bold;">{{tasks.length}}</span> of <span style="font-weight:bold;">{{tasks.length}}</span> tasks</p>
 <!-- Using the Angular Component Development Kit Table Module, and not Angular Material -->
 <table cdk-table [dataSource]="dataSource$" class="govuk-table" [multiTemplateDataRows]="true">
 
@@ -8,7 +9,7 @@
               (click)="onSortHandler(field.sortName)" >{{field.columnLabel}}</button>
     </th>
     <td class="govuk-table__cell border-none" cdk-cell *cdkCellDef="let rowData">
-      <exui-task-field [task]="rowData" [config]="field"></exui-task-field>
+      <exui-task-field-v2 [task]="rowData" [config]="field"></exui-task-field-v2>
     </td>
   </ng-container>
