--- src/work-allocation-2/release2/components/available-tasks-filter/available-tasks-filter.component.html
@@ -1,18 +1,0 @@
-<details #filterDetails class="available-tasks-filter govuk-details" data-module="govuk-details" role="group">
-  <summary class="govuk-details__summary" role="button" id="toggleFilter">
-    <span class="govuk-details__summary-text">Filter by location</span>
-  </summary>
-  <xuilib-checkbox-list #locationFilter class="available-tasks_location-filter"
-    (selectionChange)="onSelectionChange()" [options]="locations"
-    [labelFunction]="locationLabelFunction" [preselection]="preselection">
-  </xuilib-checkbox-list>
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-full">
-      <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible" />
-      <button class="govuk-button govuk-!-margin-right-1" type="submit" id="applyFilter"
-        (click)="applyFilter()">Apply</button>
-      <button class="govuk-button govuk-button--secondary" id="cancelFilter"
-        (click)="cancelFilter()">Cancel</button>
-    </div>
-  </div>
-</details>
