--- src/work-allocation-2/release2/components/available-tasks-filter/available-tasks-filter.component.scss
@@ -1,17 +1,0 @@
-:host ::ng-deep .available-tasks_location-filter-r2 {
-  .xui-checkbox-list {
-    height: 200px;
-    // border-bottom: solid 1px #b1b4b6;
-    padding-bottom: 20px;
-    .select-all {
-      margin-bottom: auto;
-    }
-    .checkbox-items {
-      margin-left: -10px;
-      padding-left: 10px;
-    }
-  }
-}
-.govuk-section-break {
-  height: auto;
-}
