--- src/work-allocation-2/release2/components/index.ts
@@ -1,13 +1,0 @@
-import { AvailableTasksFilterRelease2Component } from './available-tasks-filter/available-tasks-filter.component';
-import { CaseNameFieldComponent } from './case-name-field/case-name-field.component';
-import { TaskFieldRelease2Component } from './task-field/task-field.component';
-import { TaskNameFieldComponent } from './task-name-field/task-name-field.component';
-
-export const components: any[] = [
-  AvailableTasksFilterRelease2Component, TaskFieldRelease2Component, CaseNameFieldComponent, TaskNameFieldComponent
-];
-
-export * from './available-tasks-filter/available-tasks-filter.component';
-export * from './case-name-field/case-name-field.component';
-export * from './task-name-field/task-name-field.component';
-export * from './task-field/task-field.component';
