--- src/work-allocation-2/release2/components/task-field/task-field.component.html
@@ -1,47 +1,0 @@
-<ng-container *ngIf="config && task" [ngSwitch]="config.type" class="xui-task-field">
-  <ng-container *ngSwitchCase="fieldType.STRING">
-    {{ task[config.name] }}
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.CASE_REFERENCE_STRING">
-    {{ task[config.name] | ccdCaseReference }}
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.DATE_DUE">
-    <xuilib-due-date [dueDate]="toDate(task[config.name])"></xuilib-due-date>
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.DATE_AGE_DAYS">
-    {{ task[config.name] | daysFromToday }}
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.DATE">
-    {{ task[config.name] | date:'dd/MM/yyyy'}}
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.DATETIME">
-    {{ task[config.name] | date:'dd/MM/yyyy HH:mm'}}
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.BOOLEAN">
-    {{ task[config.name] | yesNo }}
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.INTEGER">
-    {{ task[config.name] | integer }}
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.DECIMAL_2">
-    {{ task[config.name] | twoDP }}
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.URL">
-    <exui-url-field [href]="task[config.name]"></exui-url-field>
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.IMAGE">
-    <exui-image-field [src]="task[config.name]"></exui-image-field>
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.CASE_REFERENCE">
-    <exui-case-reference-field [caseReference]="task[config.name]"></exui-case-reference-field>
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.CASE_NAME">
-    <exui-case-name-field [caseName]="task[config.name]" [caseId]="task.case_id"></exui-case-name-field>
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.TASK_NAME">
-    <exui-task-name-field [taskName]="task[config.name]" [caseId]="task.case_id"></exui-task-name-field>
-  </ng-container>
-  <ng-container *ngSwitchCase="fieldType.DERIVED_ICON">
-    <exui-derived-icon-field [task]="task" [sourceColumn]="config.sourceColumn" [matchValue]="config.matchValue"></exui-derived-icon-field>
-  </ng-container>
-</ng-container>
