--- src/work-allocation-2/release2/components/task-field/task-field.component.scss
@@ -1,7 +1,0 @@
-.xui-image-field__container {
-  height: 30px;
-  position: relative;
-  img {
-    height: 100%;
-  }
-}
