--- src/work-allocation-2/release2/components/work-allocation.components.module.ts
@@ -1,35 +1,0 @@
-import { PipesModule } from '@hmcts/ccd-case-ui-toolkit';
-import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
-import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { ExuiCommonLibModule } from '@hmcts/rpx-xui-common-lib';
-
-import * as fromComponents from '.';
-import { CaseworkerDataService, InfoMessageCommService, LocationDataService } from '../../../work-allocation/services';
-import { WorkAllocationPipesModule } from '../../../work-allocation/pipes/work-allocation.pipes.module';
-import { WorkAllocationComponentsModule } from '../../../work-allocation/components/work-allocation.components.module';
-
-// from containers
-@NgModule({
-  imports: [
-    CommonModule,
-    HttpClientModule,
-    ExuiCommonLibModule,
-    FormsModule, // TODO: Remove this as it's only needed for testing.
-    PipesModule,
-    WorkAllocationPipesModule,
-    WorkAllocationComponentsModule
-  ],
-  declarations: [
-    ...fromComponents.components
-  ],
-  providers: [CaseworkerDataService, LocationDataService, InfoMessageCommService],
-  schemas: [CUSTOM_ELEMENTS_SCHEMA],
-  exports: [
-    ...fromComponents.components
-  ]
-})
-export class WorkAllocationRelease2ComponentsModule {
-
-}
