--- src/work-allocation-2/release2/containers/available-tasks/available-tasks.component.html
@@ -1,5 +1,0 @@
-<exui-task-list-r2
-  [taskServiceConfig]="taskServiceConfig" [tasks]="tasks" [fields]="fields"
-  [sortedBy]="sortedBy" [emptyMessage]="emptyMessage"
-  (sortEvent)="onSortHandler($event)" (actionEvent)="onActionHandler($event)">
-</exui-task-list-r2>
