--- src/work-allocation-2/release2/containers/index.ts
@@ -1,9 +1,0 @@
-import { AvailableTasksRelease2Component } from './available-tasks/available-tasks.component';
-import { TaskListRelease2Component } from './task-list/task-list.component';
-
-export const containers: any[] = [
-  AvailableTasksRelease2Component, TaskListRelease2Component
-];
-
-export * from './available-tasks/available-tasks.component';
-export * from './task-list/task-list.component';
