--- src/work-allocation-2/release2/containers/task-list/task-list.component.html
@@ -1,65 +1,0 @@
-<p *ngIf="tasks">Displaying <span style="font-weight:bold;">{{tasks.length}}</span> of <span style="font-weight:bold;">{{tasks.length}}</span> tasks</p>
-<!-- Using the Angular Component Development Kit Table Module, and not Angular Material -->
-<table cdk-table [dataSource]="dataSource$" class="govuk-table" [multiTemplateDataRows]="true">
-
-  <!-- Column Headers and Cells -->
-  <ng-container *ngFor="let field of fields" [cdkColumnDef]="field.name">
-    <th class="govuk-table__header" [attr.aria-sort]="getColumnSortedSetting(field.sortName)" cdk-header-cell *cdkHeaderCellDef>
-      <button class="column-sort" *ngIf="field.columnLabel" type="button" [id]="'sort_by_' + field.sortName" data-index="0"
-              (click)="onSortHandler(field.sortName)" >{{field.columnLabel}}</button>
-    </th>
-    <td class="govuk-table__cell border-none" cdk-cell *cdkCellDef="let rowData">
-      <exui-task-field-v2 [task]="rowData" [config]="field"></exui-task-field-v2>
-    </td>
-  </ng-container>
-
-  <!-- Manage column -->
-  <ng-container *ngIf="showManage" cdkColumnDef="manage">
-    <th class="govuk-table__header" cdk-header-cell *cdkHeaderCellDef>&nbsp;</th>
-    <td cdk-cell *cdkCellDef="let rowData" class="govuk-table__cell border-none govuk-accordion__section-header">
-      <div class="govuk-accordion__section-heading">
-        <button type="button" [id]="'manage_' + rowData.id" (click)="setSelectedTask(rowData)"
-                class="govuk-accordion__section-button xui-reset"
-                [attr.aria-expanded]="isTaskSelected(rowData)">
-          Manage
-          <span class="govuk-accordion__icon" aria-hidden="true"></span>
-        </button>
-      </div>
-    </td>
-  </ng-container>
-
-  <!-- Actions for the selected row -->
-  <ng-container cdkColumnDef="actions">
-    <td cdk-cell *cdkCellDef="let rowData" [attr.colspan]="displayedColumns.length" class="cell-actions">
-      <ng-container *ngIf="isTaskSelected(rowData)">
-        <div *ngFor="let action of rowData.actions" class="task-action">
-          <a tabindex="0" (click)="onActionHandler(rowData, action)" class="action" [id]="'action_' + action.id">{{action.title}}</a>
-        </div>
-      </ng-container>
-    </td>
-  </ng-container>
-
-  <!-- No tasks to display -->
-  <ng-container cdkColumnDef="footer">
-    <td cdk-footer-cell *cdkFooterCellDef [attr.colspan]="displayedColumns.length" class="cell-footer">
-      {{emptyMessage}}
-    </td>
-  </ng-container>
-
-  <!-- Table Header Row -->
-  <tr class="govuk-table__row" cdk-header-row *cdkHeaderRowDef="displayedColumns"></tr>
-
-  <!-- Table Row -->
-  <tr class="govuk-table__row" [ngClass]="{ selected: isTaskSelected(row) }"
-      cdk-row *cdkRowDef="let row; columns: displayedColumns"></tr>
-
-  <!-- Table Actions Row-->
-  <tr class="govuk-table__row actions-row" [ngClass]="{ selected: isTaskSelected(row) }"
-      [attr.aria-hidden]="!isTaskSelected(row)"
-      cdk-row *cdkRowDef="let row; columns: ['actions']"></tr>
-
-  <!-- Table Footer Row -->
-  <tr class="govuk-table__row footer-row" [ngClass]="{ shown: tasks?.length === 0 }"
-      [attr.aria-hidden]="tasks?.length !== 0" cdk-footer-row *cdkFooterRowDef="['footer']"></tr>
-
-</table>
