--- src/work-allocation-2/release2/containers/task-list/task-list.component.scss
@@ -1,113 +1,0 @@
-.actions {
-  display: flex;
-  justify-content: flex-end;
-
-  .action {
-    padding: 2px;
-  }
-}
-/**
- * Used to override the border-bottom style being set by the GDS .govuk-table__cell
- */
-.border-none {
-  border: 0px;
-}
-
-.column-sort {
-  padding-right: 20px;
-
-  &:hover{
-    padding-right: 20px;
-  }
-
-  &:after{
-    right: 0.5px;
-  }
-  
-  &:before{
-    right: 0.5px;
-  }
-}
-
-.header-cell {
-
-  display: flex;
-
-  .arrow {
-    border: 1px solid red;
-    padding: 0 0 0 10px;
-  }
-}
-
-/**
- * The "Manage +" button should mirror the GDS accordion functionality. However,
- * it still needs to be in keeping with the styling within the table.
- */
-.govuk-accordion__section-button {
-  &.xui-reset {
-    font-size: inherit;
-    margin-right: 20px;
-    text-decoration: underline;
-    font-weight: 400;
-    &:focus {
-      text-decoration: none;
-    }
-    .govuk-accordion__icon {
-      height: 14px;
-      width: 14px;
-      right: 0;
-      top: 16px;
-      margin-top: 0;
-    }
-    &[aria-expanded=true] {
-      .govuk-accordion__icon::after {
-        display: none;
-      }
-    }
-  }
-}
-.govuk-table__row {
-  border-bottom: solid 1px #ddd;
-  th {
-    border-bottom-color: #bfc1c3;
-  }
-  .cdk-column-manage {
-    text-align: right;
-  }
-  &.actions-row {
-    display: none;
-    &.selected {
-      display: table-row;
-    }
-    .cell-actions {
-      text-align: right;
-      .task-action {
-        display: inline-block;
-        color: #1d70b8;
-        text-decoration: underline;
-        margin-right: 40px;
-        cursor: pointer;
-        &:last-child {
-          margin-right: 20px;
-        }
-      }
-    }
-  }
-  &.footer-row {
-    display: none;
-    .empty-message {
-      display: none;
-    }
-    &.shown {
-      display: table-row;
-      &.empty {
-        .empty-message {
-          display: inline;
-        }
-        .loading {
-          display: none;
-        }
-      }
-    }
-  }
-}
