+++ src/work-allocation-2/work-allocation2.module.ts
@@ -8,9 +8,7 @@ import { ExuiCommonLibModule } from '@hmcts/rpx-xui-common-lib';
 
 import { SharedModule } from '../app/shared/shared.module';
 import { WorkAllocationComponentsModule } from './components/work-allocation.components.module';
-import { WorkAllocationRelease2ComponentsModule } from './release2/components/work-allocation.components.module';
 import * as fromContainers from './containers';
-import * as fromRelease2Containers from './release2/containers';
 import { WorkAllocationFeatureToggleGuard } from './guards';
 import { CaseworkerDataService, WorkAllocationTaskService } from './services';
 import { WorkAllocationFeatureService } from './services/work-allocation-feature.service';
@@ -25,13 +23,12 @@ import { workAllocationRouting } from './work-allocation-feature.routes';
     SharedModule,
     MatDialogModule,
     WorkAllocationComponentsModule,
-    WorkAllocationRelease2ComponentsModule,
     PipesModule,
     workAllocationRouting,
     CdkTableModule,
     ExuiCommonLibModule,
   ],
-  declarations: [...fromContainers.containers, ...fromRelease2Containers.containers],
+  declarations: [...fromContainers.containers],
   providers: [
     WorkAllocationTaskService,
     WorkAllocationFeatureToggleGuard,
